Handles the START response
Closes #8

Replaces the previous started boolean with an enum of possible states.

Only set the status to started after receiving the START response.
Previously we assumed we were started after we sent the initial START
message.

This makes room for a stream so apps can listen to the status changes,
but this does not include the stream.

Add unit tests for verifying status set correctly.